create an app component to house the full module, put overlay component in its own file, render overlay via portal so that it appears on a click event
