+++ public/index.html
@@ -10,6 +10,7 @@
 
   <body>
     <div id="app"></div>
+    <div id="container"></div>
     <script src="./bundle.js"></script>
   </body>
 </html>
